+++ xivo_confd/plugins/skill/service.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+# Copyright 2018 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0+
+
+from xivo_dao.resources.skill import dao as skill_dao
+
+from xivo_confd.helpers.resource import CRUDService
+
+from .notifier import build_notifier
+from .validator import build_validator
+
+
+def build_service():
+    return CRUDService(skill_dao, build_validator(), build_notifier())